+++ ui/src/plugins/argocd-plugin/argocd-plugin.tsx
@@ -1,0 +1,11 @@
+import { PluginsInstallation } from '../atoms/plugin-interfaces';
+
+import promotionDeepLinkPlugin from './deep-link/promotion';
+
+const plugin: PluginsInstallation = {
+  DeepLinkPlugin: {
+    Promotion: promotionDeepLinkPlugin
+  }
+};
+
+export default plugin;
